requirements/local.txt
Django==1.7.4
djangorestframework==3.0.5
oauthlib==0.7.2
PyJWT==0.4.3
python-openid==2.2.5
python-social-auth==0.2.1
requests==2.5.1
requests-oauthlib==0.4.2
six==1.9.0
twython==3.2.0